fix: address review follow-ups for struct out-buffers

+++ tests/e2e/cases/09-stateful/enrollment/enrollment_disproved.cpp
@@ -17,5 +17,5 @@ std::uint32_t enroll_key(EnrollmentKey* k) noexcept {
     std::uint8_t engaged = k->engaged;
     k->createdAt = 42;
     k->engaged = engaged ^ 1; // BUG: model keeps engaged unchanged
-    return engaged;
+    return static_cast<std::uint32_t>(engaged);
 }
